[hive] Refactor to simplify  and make logic more clear
